refactor test_read.py
